isolate buffer logic
